Cleanup stuff

+++ AudioKit/Common/Nodes/Playback/Samplers/Wave Table/AKWaveTableAudioUnit.mm
@@ -44,7 +44,7 @@
 - (void)setLoop:(BOOL)loopOnOff {
     _kernel.setLoop(loopOnOff);
 }
-- (void)setRate:(double)rate {
+- (void)setRate:(float)rate {
     _kernel.setRate(rate);
 }
 - (void)setVolume:(float)volume {
@@ -59,7 +59,7 @@
 - (int)size {
     return _kernel.ftbl_size;
 }
-- (double)position {
+- (float)position {
     float normalized = (_kernel.position - _kernel.startPointViaRate()) / (_kernel.endPointViaRate() - _kernel.startPointViaRate());
     return _kernel.rate > 0 ? normalized : 1 - normalized;
 }
